Move calibrated_salinity.h into the lib/utils directory.
